requirements.txt
absl-py==0.10.0
astor==0.8.1
astunparse==1.6.3
attrs==20.3.0
boto3==1.14.56
botocore==1.17.56
cached-property==1.5.2
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
clearml==0.17.5
dataclasses==0.8
dill==0.3.3
docutils==0.15.2
flatbuffers==1.12
furl==2.1.0
future==0.18.2
gast==0.3.3
google-auth==1.27.1
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.32.0
h5py==2.10.0
humanfriendly==9.1
idna==2.10
importlib-metadata==3.7.3
importlib-resources==5.1.2
jmespath==0.10.0
jsonschema==3.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.4
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
orderedmultidict==1.0.1
pathlib2==2.3.5
Pillow==8.1.2
promise==2.3
protobuf==3.15.6
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.3.3
six==1.15.0
tensorboard
tensorboard-plugin-wit
tensorflow-datasets
tensorflow-estimator
tensorflow-metadata
termcolor==1.1.0
tqdm==4.59.0
typing-extensions==3.7.4.3
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.1
